+++ synapse-aws-sqs/src/test/java/de/otto/synapse/acceptance/SqsAcceptanceTest.java
@@ -26,7 +26,7 @@ import java.util.concurrent.atomic.AtomicReference;
 
 import static de.otto.synapse.configuration.sqs.SqsTestConfiguration.SQS_INTEGRATION_TEST_CHANNEL;
 import static de.otto.synapse.message.DefaultHeaderAttr.*;
-import static de.otto.synapse.message.Header.requestHeader;
+import static de.otto.synapse.message.Header.of;
 import static de.otto.synapse.message.Message.message;
 import static java.time.Instant.ofEpochSecond;
 import static java.util.concurrent.TimeUnit.SECONDS;
@@ -100,7 +100,7 @@ public class SqsAcceptanceTest {
 
     @Test
     public void shouldSendAndReceiveCustomSqsMessageHeaders() {
-        final Header header = requestHeader(ImmutableMap.of(
+        final Header header = of(ImmutableMap.of(
                 "string", "some value",
                 "timestamp", ofEpochSecond(42).toString())
         );