+++ synapse-core/src/test/java/de/otto/synapse/consumer/StatefulMessageConsumerTest.java
@@ -8,7 +8,7 @@ import org.mockito.Mock;
 import org.mockito.junit.MockitoJUnitRunner;
 
 import static de.otto.synapse.channel.ShardPosition.fromPosition;
-import static de.otto.synapse.message.Header.responseHeader;
+import static de.otto.synapse.message.Header.of;
 import static org.mockito.Mockito.verify;
 
 @RunWith(MockitoJUnitRunner.class)
@@ -25,7 +25,7 @@ public class StatefulMessageConsumerTest {
         //when
         consumer.accept(Message.message(
                 "someKey",
-                responseHeader(fromPosition("some-shard", "someSeqNumber")),
+                of(fromPosition("some-shard", "someSeqNumber")),
                 "12345"
         ));
 
@@ -41,7 +41,7 @@ public class StatefulMessageConsumerTest {
         //when
         consumer.accept(Message.message(
                 "someKey",
-                responseHeader(fromPosition("some-shard", "someSeqNumber")),
+                of(fromPosition("some-shard", "someSeqNumber")),
                 "12345"
         ));
 
@@ -57,7 +57,7 @@ public class StatefulMessageConsumerTest {
         //when
         consumer.accept(Message.message(
                 "someKey",
-                responseHeader(fromPosition("some-shard", "someSeqNumber")),
+                of(fromPosition("some-shard", "someSeqNumber")),
                 "12345"
         ));
 
@@ -73,7 +73,7 @@ public class StatefulMessageConsumerTest {
         //when
         consumer.accept(Message.message(
                 "someKey",
-                responseHeader(fromPosition("some-shard", "someSeqNumber")),
+                of(fromPosition("some-shard", "someSeqNumber")),
                 null
         ));