+++ synapse-core/src/test/java/de/otto/synapse/eventsource/DefaultEventSourceTest.java
@@ -5,6 +5,7 @@ import de.otto.synapse.channel.ChannelPosition;
 import de.otto.synapse.consumer.MessageDispatcher;
 import de.otto.synapse.endpoint.InterceptorChain;
 import de.otto.synapse.endpoint.receiver.MessageLogReceiverEndpoint;
+import de.otto.synapse.message.Header;
 import de.otto.synapse.message.Key;
 import de.otto.synapse.message.Message;
 import de.otto.synapse.messagestore.MessageStore;
@@ -18,7 +19,6 @@ import static de.otto.synapse.channel.ChannelPosition.channelPosition;
 import static de.otto.synapse.channel.ChannelPosition.fromHorizon;
 import static de.otto.synapse.channel.ShardPosition.fromPosition;
 import static de.otto.synapse.channel.StopCondition.shutdown;
-import static de.otto.synapse.message.Header.responseHeader;
 import static de.otto.synapse.message.Message.message;
 import static de.otto.synapse.messagestore.MessageStores.emptyMessageStore;
 import static java.util.concurrent.CompletableFuture.completedFuture;
@@ -62,7 +62,7 @@ public class DefaultEventSourceTest {
         // given
         // and some message store having a single message
         final MessageStore messageStore = mock(MessageStore.class);
-        when(messageStore.stream()).thenReturn(Stream.of(message(Key.of("1"), responseHeader(null), null)));
+        when(messageStore.stream()).thenReturn(Stream.of(message(Key.of("1"), null)));
         when(messageStore.getLatestChannelPosition()).thenReturn(fromHorizon());
         // and some MessageLogReceiverEndpoint with an InterceptorChain:
         final InterceptorChain interceptorChain = mock(InterceptorChain.class);
@@ -78,7 +78,7 @@ public class DefaultEventSourceTest {
 
         // then
         verify(interceptorChain).intercept(
-                message(Key.of("1"), responseHeader(null), null)
+                message(Key.of("1"), null)
         );
     }
 
@@ -87,7 +87,7 @@ public class DefaultEventSourceTest {
         // given
         // and some message store having a single message
         final MessageStore messageStore = mock(MessageStore.class);
-        when(messageStore.stream()).thenReturn(Stream.of(message(Key.of("1"), responseHeader(null), null)));
+        when(messageStore.stream()).thenReturn(Stream.of(message(Key.of("1"), Header.of(), null)));
         when(messageStore.getLatestChannelPosition()).thenReturn(fromHorizon());
         // and some InterceptorChain that is dropping messages:
         final InterceptorChain interceptorChain = new InterceptorChain(ImmutableList.of((m)->null));
@@ -112,7 +112,7 @@ public class DefaultEventSourceTest {
         // given
         // and some message store having a single message
         final MessageStore messageStore = mock(MessageStore.class);
-        when(messageStore.stream()).thenReturn(Stream.of(message(Key.of("1"), responseHeader(null), null)));
+        when(messageStore.stream()).thenReturn(Stream.of(message(Key.of("1"),null)));
         when(messageStore.getLatestChannelPosition()).thenReturn(fromHorizon());
         // and some MessageLogReceiverEndpoint with our InterceptorChain:
         final MessageLogReceiverEndpoint messageLog = mock(MessageLogReceiverEndpoint.class);
@@ -127,7 +127,7 @@ public class DefaultEventSourceTest {
         eventSource.consume().get();
 
         // then
-        verify(messageDispatcher).accept(message(Key.of("1"), responseHeader(null), null));
+        verify(messageDispatcher).accept(message(Key.of("1"), null));
     }
 
     @Test