+++ synapse-core/src/test/java/de/otto/synapse/messagestore/NonCompactingMessageStoreTest.java
@@ -12,7 +12,7 @@ import java.util.concurrent.ExecutorService;
 import java.util.function.Supplier;
 
 import static de.otto.synapse.channel.ShardPosition.fromPosition;
-import static de.otto.synapse.message.Header.responseHeader;
+import static de.otto.synapse.message.Header.of;
 import static de.otto.synapse.message.Message.message;
 import static java.lang.String.valueOf;
 import static java.util.Arrays.asList;
@@ -50,7 +50,7 @@ public class NonCompactingMessageStoreTest {
             final String shardId = valueOf(shard);
             completion[shard] = CompletableFuture.runAsync(() -> {
                 for (int pos = 0; pos < 10000; ++pos) {
-                    messageStore.add(message(Key.of(valueOf(pos)), responseHeader(fromPosition("shard-" + shardId, valueOf(pos))), "some payload"));
+                    messageStore.add(message(Key.of(valueOf(pos)), of(fromPosition("shard-" + shardId, valueOf(pos))), "some payload"));
                     assertThat(messageStore.getLatestChannelPosition().shard("shard-" + shardId).startFrom(), is(StartFrom.POSITION));
                     assertThat(messageStore.getLatestChannelPosition().shard("shard-" + shardId).position(), is(valueOf(pos)));
                 }